refactor(frontend): restructure GldtStake components
