+++ src/frontend/src/icp/components/stake/gldt/GldtStakeEarnCard.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
 	import { isNullish, nonNullish } from '@dfinity/utils';
-	import { setContext } from 'svelte';
-	import GldtStakeApyContext from '$icp/components/stake/gldt/GldtStakeApyContext.svelte';
 	import {
 		enabledIcrcTokens,
 		icrcTokens,
@@ -9,11 +7,6 @@
 		icrcCustomTokensInitialized
 	} from '$icp/derived/icrc.derived';
 	import { loadCustomTokens } from '$icp/services/icrc.services';
-	import {
-		GLDT_STAKE_APY_CONTEXT_KEY,
-		type GldtStakeApyContext as GldtStakeApyContextType,
-		initGldtStakeApyStore
-	} from '$icp/stores/gldt-stake-apy.store';
 	import { setCustomToken } from '$icp-eth/services/custom-token.services';
 	import { isGLDTToken } from '$icp-eth/utils/token.utils';
 	import StakeContentCard from '$lib/components/stake/StakeContentCard.svelte';
@@ -35,10 +28,6 @@
 	import { formatCurrency, formatToken } from '$lib/utils/format.utils';
 	import { replacePlaceholders } from '$lib/utils/i18n.utils';
 	import { calculateTokenUsdAmount, getTokenDisplaySymbol } from '$lib/utils/token.utils';
-
-	setContext<GldtStakeApyContextType>(GLDT_STAKE_APY_CONTEXT_KEY, {
-		store: initGldtStakeApyStore()
-	});
 
 	let gldtToken = $derived($enabledIcrcTokens.find((token) => isGLDTToken(token)));
 
@@ -76,57 +65,55 @@
 	};
 </script>
 
-<GldtStakeApyContext>
-	<StakeContentCard>
-		{#snippet content()}
-			{#if nonNullish(gldtToken)}
-				<span class="font-bold">
-					{formatCurrency({
-						value: gldtTokenUsdBalance,
-						currency: $currentCurrency,
-						exchangeRate: $currencyExchangeStore,
-						language: $currentLanguage
-					})}
-				</span>
+<StakeContentCard>
+	{#snippet content()}
+		{#if nonNullish(gldtToken)}
+			<span class="font-bold">
+				{formatCurrency({
+					value: gldtTokenUsdBalance,
+					currency: $currentCurrency,
+					exchangeRate: $currencyExchangeStore,
+					language: $currentLanguage
+				})}
+			</span>
 
-				<span class="text-tertiary">
-					{formatToken({
-						value: gldtTokenBalance,
-						unitName: gldtToken.decimals
-					})}
-					{gldtTokenSymbol}
-				</span>
-			{:else if $icrcCustomTokensInitialized}
-				<span class="text-tertiary">
-					{replacePlaceholders($i18n.stake.text.enable_token_text, {
-						$token_symbol: gldtTokenSymbol
-					})}
-				</span>
-			{/if}
-		{/snippet}
+			<span class="text-tertiary">
+				{formatToken({
+					value: gldtTokenBalance,
+					unitName: gldtToken.decimals
+				})}
+				{gldtTokenSymbol}
+			</span>
+		{:else if $icrcCustomTokensInitialized}
+			<span class="text-tertiary">
+				{replacePlaceholders($i18n.stake.text.enable_token_text, {
+					$token_symbol: gldtTokenSymbol
+				})}
+			</span>
+		{/if}
+	{/snippet}
 
-		{#snippet buttons()}
-			{#if nonNullish(gldtToken)}
-				<ButtonWithModal isOpen={$modalGldtStake} onOpen={modalStore.openGldtStake}>
-					{#snippet button(onclick)}
-						<Button disabled={gldtTokenBalance === ZERO} fullWidth {onclick}>
-							{replacePlaceholders($i18n.stake.text.stake, {
-								$token_symbol: gldtTokenSymbol
-							})}
-						</Button>
-					{/snippet}
+	{#snippet buttons()}
+		{#if nonNullish(gldtToken)}
+			<ButtonWithModal isOpen={$modalGldtStake} onOpen={modalStore.openGldtStake}>
+				{#snippet button(onclick)}
+					<Button disabled={gldtTokenBalance === ZERO} fullWidth {onclick}>
+						{replacePlaceholders($i18n.stake.text.stake, {
+							$token_symbol: gldtTokenSymbol
+						})}
+					</Button>
+				{/snippet}
 
-					{#snippet modal()}
-						<StakeModal token={gldtToken} />
-					{/snippet}
-				</ButtonWithModal>
-			{:else}
-				<Button fullWidth loading={$icrcCustomTokensNotInitialized} onclick={enableStakingToken}>
-					{replacePlaceholders($i18n.stake.text.enable_token_button, {
-						$token_symbol: gldtTokenSymbol
-					})}
-				</Button>
-			{/if}
-		{/snippet}
-	</StakeContentCard>
-</GldtStakeApyContext>
+				{#snippet modal()}
+					<StakeModal token={gldtToken} />
+				{/snippet}
+			</ButtonWithModal>
+		{:else}
+			<Button fullWidth loading={$icrcCustomTokensNotInitialized} onclick={enableStakingToken}>
+				{replacePlaceholders($i18n.stake.text.enable_token_button, {
+					$token_symbol: gldtTokenSymbol
+				})}
+			</Button>
+		{/if}
+	{/snippet}
+</StakeContentCard>
