+++ src/frontend/src/tests/icp/components/stake/gldt/GldtStakeEarnCard.spec.ts
@@ -1,12 +1,12 @@
 import { GLDT_LEDGER_CANISTER_ID } from '$env/networks/networks.icrc.env';
-import GldtStakeContentCard from '$icp/components/stake/gldt/GldtStakeContentCard.svelte';
+import GldtStakeEarnCard from '$icp/components/stake/gldt/GldtStakeEarnCard.svelte';
 import { icrcCustomTokensStore } from '$icp/stores/icrc-custom-tokens.store';
 import { replacePlaceholders } from '$lib/utils/i18n.utils';
 import en from '$tests/mocks/i18n.mock';
 import { mockIcrcCustomToken } from '$tests/mocks/icrc-custom-tokens.mock';
 import { render } from '@testing-library/svelte';
 
-describe('GldtStakeContentCard', () => {
+describe('GldtStakeEarnCard', () => {
 	beforeEach(() => {
 		icrcCustomTokensStore.resetAll();
 	});
@@ -17,7 +17,7 @@ describe('GldtStakeContentCard', () => {
 			certified: false
 		});
 
-		const { container } = render(GldtStakeContentCard);
+		const { container } = render(GldtStakeEarnCard);
 
 		expect(container).toHaveTextContent(
 			replacePlaceholders(en.stake.text.enable_token_button, { $token_symbol: 'GLDT' })
@@ -39,7 +39,7 @@ describe('GldtStakeContentCard', () => {
 			certified: false
 		});
 
-		const { container } = render(GldtStakeContentCard);
+		const { container } = render(GldtStakeEarnCard);
 
 		expect(container).toHaveTextContent(
 			replacePlaceholders(en.stake.text.stake, { $token_symbol: 'GLDT' })
